fix windows build error

+++ RHI/D3d/D3d12GraphicsManager.hpp
@@ -28,15 +28,15 @@ class D3d12GraphicsManager : public GraphicsManager {
 
     void DrawBatch(const Frame& frame) final;
 
-    int32_t GenerateCubeShadowMapArray(const uint32_t width,
+    intptr_t GenerateCubeShadowMapArray(const uint32_t width,
                                        const uint32_t height,
                                        const uint32_t count) final;
-    int32_t GenerateShadowMapArray(const uint32_t width, const uint32_t height,
+    intptr_t GenerateShadowMapArray(const uint32_t width, const uint32_t height,
                                    const uint32_t count) final;
-    void BeginShadowMap(const int32_t light_index, const int32_t shadowmap,
+    void BeginShadowMap(const int32_t light_index, const intptr_t shadowmap,
                         const uint32_t width, const uint32_t height,
                         const int32_t layer_index, const Frame& frame) final;
-    void EndShadowMap(const int32_t shadowmap, const int32_t layer_index) final;
+    void EndShadowMap(const intptr_t shadowmap, const int32_t layer_index) final;
     void SetShadowMaps(const Frame& frame) final;
     void CreateTexture(SceneObjectTexture& texture) final;
     void ReleaseTexture(intptr_t texture) final;
